sm6150-common: overlay: Don't pin launcher app in memory
Similar to what we did for the camera app, unpin the launcher app from
memory as well. While the default launcher (Launcher3) isn't
particularly big, it doesn't make much sense to pin because the launcher
does not typically load new resources much. Most of its resources should
already be loaded in memory after it starts, so pinning the APK is
redundant.

Signed-off-by: drkphnx <dark.phnx12@gmail.com>
Change-Id: I233e34ac4e9aa053baf6634fa3f9b4a508656334
Signed-off-by: venom-stark <sachidanandabehera562@gmail.com>