sm6150-common: Label android.hardware.drm@1.4-service.clearkey
I hwservicemanager: Since android.hardware.drm@1.0::IDrmFactory/clearkey is not registered, trying to start it as a lazy HAL.
I HidlServiceManagement: getService: Trying again for android.hardware.drm@1.0::IDrmFactory/clearkey...
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.drm@1.0::IDrmFactory/clearkey": error code: 0x20
I hwservicemanager: Tried to start android.hardware.drm@1.0::IDrmFactory/clearkey as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Signed-off-by: Anush02198 <Anush.4376@gmail.com>
Signed-off-by: venom-stark <sachidanandabehera562@gmail.com>